Added initial support for SLURM jobs
The tool can build a basic script to be run by sbatch (to be tested).
To avoid code redundancy, `build_qsub_cmd` was split into two parts,
- `build_qsub_head` generates the qsub specific commands.
- `build_bash_cmd` generates the BASH commands.
The function to build extra commands was not anymore necessary and
integrated into the header-specific function.
A new function, `build_sbatch_head`, builds the header for a
SLURM-compatible job, in a way symmetric to `build_qsub_head`.