build_cluster supports gdv.jXXp-<dates> formats
Intermediate versions of GDV are specified with the + sign and multiple
versions can be produced between two "snapshots".  To distinguish them
and avoid overlaps, it may be convenient to add a date to the file, for
instance gdv.jXXp-YYMMDD.  This format is now supported as a valid
archive file format.  The script can support 6 or 8 digits for the date.
This is only supported for such intermediate versions, since snapshots
versions are unique.